doc search: support search specific to a language family
Also, revise the search-page interface a little: use icons for help
and setting options, move the context (i.e., query prefix) info to the
top, and refine the query-prefix panel. Sort indices are used to order
exact matches.

Add a configuration key for the main language family of an
installation. This language family is used when reporting search
results as the default family, and search results for that family are
not cluttered with an extra label to note the family. If a search is
made specific to a family, the results similarly omit an extra label.

Support external declaration of a documentation's language family
through `scribblings` in "info.rkt", and allow a document (or Scribble
dialect) to declare a defualt language family to be used if
`scribblings` does not specify one. Meanwhile, individual index
entries can have their own language-family specifications to override
a document-wide default.