Chez Scheme: add `stdbool` foreign type
The `stdbool` type corresponds to `bool` as defined by the host
machine's "stdbool.h", which is 1 byte for all currently supported
platforms. That's different than `boolean`, which instead reflects a
traditional C encoding of booleans as `int`, which is 4 bytes for all
currently supported platforms.

The "stdbool.h" header is standard as of C99, while Chez Scheme's
implementation generally sticks to C89. But the header needed only for
a test file, and it tends to be available in C89 enviornments, too.