error messages with values: add multi-line convention
Previously, error messages were constructed with the assumption that
the string form of each value in the message fits on a single line. To
support values that print with multiple lines, start them on a new
line and add indentation before each line.

It was formerly possible to achieve that output in most cases by
having the string form of a value (or an unquoted-printing string)
start with a newline and have each line's indentation built in. To
accomodate those potential existing uses, do not perform an
auto-indenting conversion if the string form of a value starts with a
newline. Starting with a newline and building indentation into the
string does not quite work right in cases where an error message shows
an indented sequence of value, though, so that's why it's better to
build the convention into message-construction functions, instead of
in an `error-value->string-handler` function.