raco setup: avoid discarding moved libraries that appear elsewhere
When `raco setup` considers whether it should move a shared library
into place based on whether the same library exists already in the
search path, it could end up discarding a library that was put in
place in "move" mode, which is the case for shared libraries that were
put in place for a distribution. There was a specific attempt within
`raco setup` to handle "move" mode in that situation, but it was not
the right choice.

The test for this change is in the `distro-build-test` package's
`unix-installer-test.rkt`.

Closes #5100