allow `make-parameter` as cross-phase persistent
Allow non-system parameters to cross phases by allowing
`make-parameter` in a cross-phase persistent mode. Use that capability
for the `syntax-parse` configuration parameter.

Racket has various built-in parameters that span phases. Those
parameters provide a way to escape from the expander's phase
separation, so I've always hoped/wished that they could go away. But
their phase-spanning behavior is handy and maybe unavoidable for
system-wide configuration, such as adapting the environment to a
language's conventions. That kind of capability should not be
restricted to built-in parameters. Absent a better idea, I'm finally
ready to yield on allowing `make-parameter` in a cross-phase
persistent module.