limit read-bytes request size for custom input port
When the input port is created with `make-input-port`, then a read
operation makes a new buffer for each read request sent to the port,
which is attempt to help isolate the reading program and port
implementations. When `read-bytes` is called with a large number
(e.g., megabytes) or `read-bytes!` is called with a large buffer, a
new buffer of that size was created for every request, even though the
read is likely to fill only a small fraction of it. That allocation
strategy is especially bad for `read-bytes` and an SSL port, making
the read take quadratic time for a large `read-bytes` request.

The intermediate buffer is now limited to 8k. It's possible that this
limit will get in the way of a port that can deliver more bytes at
once, but those uses seem much more rare than requesting many bytes
from an SSL input port.