json: add `#:replace-malformed-surrogate?` argument
As specified in the the JSON grammar 16 bit unicode escapes are allowed with
`\uHHHH`.  To write a 32 bit code point the specification requires the escapes
to be written with surrogate pairs ala UTF-16.  When these surrogate pairs are
malformed the previous implementation would raise an exception.  This patch
changes the behavior by giving an option to raise an exception or replace the
malformed bytes with the unicode replacement character. The replacement behavior
implemented here is consistent with other JSON processing tools.