raco demod: improve pruning of unused assignments
The change reduces the demodularized form of a Rhombus "Hello World"
by about 50% by dropping a knot-tying `set!` where the knot is
otherwise unreferenced.

Also, report an error when an excluded or included module is not a
dependency of the input module. This change is not backward
compatible, but the `compiler/demod` interface is new and used
narrowly enough, so far, that the check seems worthwhile.