CS: repair FFI call cooperation with scheduler's atomic mode
An FFI call at the Rhombus layer with `#:in-original-place?` was
setting the atomic counter directly, instead of using `end-atomic` to
end atomic mode at the scheduler layer. As a result, callbacks to be
run at the end of atomic mode could get dropped. A dropped callback
would normally trigger a thread swap as a thread exits atomic mode
(because its quantum expired while it was in atomic mode), so missing
the swap could keep a thread scheduled too long.

Recent FFI improvements exposed the problem, because it caused the
"ffi-orig-place.rkt" benchmark to often run with just the right
schedule to have a thread stay swapped in, and so the benchmark would
not finish.

Also, fix some tests that wer not running as intended.