optimize `(apply string-append ....)`
When appending many strings, the problem with `(apply string-append strs)`
is that the `strs` list is unpacked as arguments and then repacked as
a new rest-arg list. Also, the underlying `string-append` in Chez Scheme
creates a continuation proportional to the size of the list in that case.
To avoid those problems, optimize `(apply string-append strs)` to
`(apply-string-append 0 strs)`, where `apply-string-append` is an
internal function. More generally, `(apply string-append str ... strs)`
becomes `(apply-string-append N (list* str ... strs))` for N `str`s.

Perform the same conversion for `string-append-immutable` and
`bytes-append`, too.

This optimization is applied by schemify, so it's only in CS. The
internal functions are implemented for BC to work with the schemified
expander, but BC doesn't try to perform the `apply string-append`
conversion (although it could).

Fixes #5189