expander: repair inlining of protected exports from `#%core`
The internal `#%core` module exports a few bindings as protected,
including `local-expand`. If these bindings are eventually inlined
into an untrusted context via a trusted intermediary, there ends up
being a check on the declaration-time code inspector of `#%core`, and
that inspector was not properly reported when queried for cross-module
inlining.