Windows CS zuofile: trigger relinking of DLL
Add a dependency to the rule for linking "libracket.dll".

In principle, every zuofile rule should depend on its enclosing
zuofile and and referenced libraries --- but that's a coarse form of
tracking, it creates too much work in practice, and relevant zuofile
changes are usually reflected by depencency changes, anyway. It
happens that the MinGW linking change isn't implicitly reflected by a
dependency, though, so introduce an explicit dependency.