pairs/lists: Introduce 4 additional algorithms
Adds some more algorithms that are commonly found in other functional programming languages.
* slice-by
* windows
* running-foldl
* running-foldr

The fold variants are variadic to keep with the conventions of the current fold algorithms.
Furthermore, I chose to make it so that windows does not accept partial windows since this is how the
algorithm is normally implemented.

Issue: https://github.com/racket/racket/issues/5210#issue-2852668731