add `bitwise-first-bit-set`
The only use of this function I find in the wild is in "math-lib", but
it's also the one function that makes "math-lib" depend on the
"r6rs-lib" package. Since the function is already implemented in Chez
Scheme, export it to Racket.

The "r6rs-lib" library will change to re-export
`bitwise-first-bit-set`, and that will reduce the potential for import
conflicts even further.