pkg: for file URL without associated ".CHECKSUM", download to infer
Previously, for a file URL without an acompanying ".CHECKSUM" file,
the checksum was simply treated as unknown, and so `raco update` for
the package would never try to update. The change here means
downloading the package to compute a checksum, if necessary. A cache
mapping etags to checksum provides a shortcut to downloads if the
server supports etags and reports no changes based on an etag.

Related to #5232