enable `SO_KEEPALIVE` on TCP connections
Setting `SO_KEEPALIVE` causes a TCP connection that is just waiting
for input (as opposed to sending or waiting for an ACK for sent data)
to time out if the oter end doesn't respond to a periodic ping. The
ping tends to be on the order of 2 hours, depending on how the OS is
configured; so, `SO_KEEPALIVE` does not affect typical connections,
but it can aid recovery from extreme stalls.