"compiled/ephemeral": use for `compiler/demod`, auto-prune for pkg
The "compiled/demod" directory created by `compile/demod` effectively
embeds absolute paths, which makes it contain information specific to
a build machine. Fortuenately, this information is not needed to check
whether a compilation is up-to-date, only to speed up recompilation.
So, it can be dropped when creating a built package, which should not
depend on the build environment.

Each package that needs this could explicitly exclude "compiler/demod"
in "info.rkt" as a pruned path, but that's tedious and fragile. It's
easiest and more consistent for `compiler/demod` to indicate the
ephemeral nature of the "compiled/demod" content.

Toward that end, "compiled/ephemeral" is now an automatically pruned
directory for `raco pkg create`, and `compiler/demo` now uses
"compiled/ephemeral/demod" instead of "compiled/demod".