BC: repair JIT bug related to `case-lambda`
When a `case-lambda` was called with arguments that are not already on
the runstack, when the number of arguments was more than a fewm and
when the specific case within the `case-lambda` had not yet been
JITted, then the `max_let_depth` value checked before calling the
function was not yet ready enough. As a result, it was possible for
copying arguments into the runstack to create a stack overflow.

This bug is about as old as the jIT. It's difficult to trigger,
because arguments are rarely delivered not on the runstack, and the
current runstack has to be almost full at the first call of a
`case-lambda` with arguments delivered in the unusual way.