repair cross-read and cross-write of ".zo"
Previously, using "chez-scheme" as the ".zo"-level tag allowed ".zo"
content for a 64-bit platform to be read on a 64-bit platforms, but it
would fail for a platform of a different size. Commit 65864a2 changed
the ".zo" tag to be specific to a machine type, which exposed the
problem that `raco exe` expected to be able to read ".zo" files for
a different platform. This commit adjusts the ".zo" encoding to make
that work, which restores the ability for platforms that use the same
word size and repairs it for different word sizes.