expander: fix variable distnctness in extraction simplifier
A simplificstion pass in extracting an expander into a single linklet
relies on a property of expanded linklets, which is that each local
binding has a distinct symbolic name. But that property did not hold
for the flattened combination of linklets that is generated in an
earlier step. So, information discovered about a binding could be
incorrectly attributed to another binding, too.