struct: add `#:properties prop-alist-expr` option
Add a #:properties option to struct that takes an association
list (same type as make-struct-type).

This makes struct a better target for macros, since the
calculation of properties can be separated from the rest of the
struct declaration. Multiple property values can be calculated in
a single expression position, making it easier for property
values to share intermediate calculations and syntax bindings.