native-libs: patch for Cairo PDF-font-substitution issue
This seems unlikely to be the right repair, because the goal is to
make the current Cairo build behave more like v1.16, and the patched
line was the same in v1.16. But this repair seems to be sufficient for
the immediate problem, and it should be harmless otherwise. The
`racket/draw` library does not depend on this repair, because it
handles font substitution itself.

Closes #5269