configure: add `--enable-missingpkgs` and `--enable-pkgs=<pkgs>`
These flags are intended for use with a source distribution and
building on Mac OS or Windows, where extra packages need to be
installed before `raco setup` for the build to work. For Windows,
`winfig.bat` accepts `/addmissingpkgs`. These additions change the
recommendation for building from a source distribution from "don't do
that" to "use `--enable-missingpkgs` or `/addmissingpkgs`".

An `--enable-catalog=<url>` option is also supported to specify the
catalog for installing extra packages.