configure: infer `--enable-missingpkgs` default
When creating a source distribution, add a "src/source-dist.txt" file
to be recongized by the built scripts to select a better default for
installing missing packages. That way, someone trying to build on Mac
OS or Windows does not need to know about `--enable-missingpkgs`
or `/addmissingpkgs`, which simplifies the instructions.

A drawback is that `make` or `make install` might access the network
even though the user did not specifically ask it to. That has become
common with other packages, though, and it's restructed here to the
case of starting with a source distribution and building on Windows,
Mac OS (without `--enable-xonx` and without `--enable-nonatipkg`), or
natipkg x86_64 Linux (with `--enable-natipkg`) --- all of which are
rare build modes.