net,http-client: handle RST during body read
An RST packet received while reading a response with a Content-Length
header would cause a thread reading from the body output port to block
indefinitely as the piping thread would crash without closing the
output side of the pipe. The same situation for a chunked transfer
would lead to a slightly better outcome, namely that the output side
of the pipe would get closed, so the reader wouldn't block
indefinitely, but it would end up receiving partial data with no
notice that something went wrong in the process.

This change wraps the input side of these pipes into a custom input
port that propagates any errors from the piping thread to the thread
reading the input port.

Unfortunately, it's not possible to trigger a RST using portable
Racket code, so the added test only works on POSIX-compatible
platforms and relies on the test machine having at least a C compiler.