ffi/unsafe/atomic: add `unsafe-{start,end}-uninterruptable`
"Atomic" means "do not let other threads run". "Uninterruptable" means
"other threads potentially run, but they should interrupt this thread,
so it should continue until at least the end of uninterrupted mode".
"Atomic" implies "uninterruptable", and atomic mode implements
uninterruptable mode for Racket threads. For futures running
concurrently to a Racket thread, however, "uninterruptable" is
implemented so that it does not block the future.

A limitation is that a future remains uninterruptable only as long as
it does not perform any action that blocks the future. Putting
uninterruptable mode to use, then, requires some knowledge of Racket
internals, similar to some other unsafe operations.

Related to #5297