add parallel threads via a `#:pool` argument to `thread`
The thread function changes to

```
 (thread thunk
        [#:pool pool
         #:keep-results? keep-results?]) -> thread?
   pool: (or/c #f 'own parallel-thread-pool?) = #f
   keep: (or/c #f 'results) = #f
```

The default `#f` for `#:pool` creates a coroutine thread, like before.
Supplying `'own` for `#:pool` creates a parallel thread that uses its
own OS thread. Supplying a thread pool creates a parallel thread that
is backed with the pool's OS threads (i.e., OS threads that are
separate from the coroutine OS thread, but shared by threads in the
pool—still parallel to each other, too). If `keep` is `'results`, then
thread-wait will report the result of thunk; that's an extension for
coroutine threads as well as parallel threads.

The implementation is based on the existing futures layer, and so
performance improvements can be expected when large regions a thread's
work would run without blocking in a future. The difference is that
when a blocking operation is encountered in a parallel thread, it is
kicked over to a coroutine shadow thread; once the blocking operation
completes[*], the continuation can be moved afterward back to the
parallel thread's future.

[*] The notion of a blocking operation "completing" falls out
naturally in the way most blocking operations are implemented.
Consistent with the existing terminology for futures, we could say
that all blocking operations for futures turn into synchronizing
operations for parallel threads.

From the perspective of a library that aims to be thread-safe, this
change doesn't introduce much new machine-level concurrency compared
to what was possible before via futures. It does make machine-level
concurrency more accessible, so we'll likely find places that weren't
synchronized correctly before. Also, it's only mostly the same,
because a library could conceivably depend on a blocking operation to
ensure sequential consistency afterward—but that seems unlikely
compared to the probably of plain old lock bugs.

Note that if a Racket program uses semaphores, then it's only the
semaphore wait or post that is atomic, and the part in between can run
in parallel. That's almost certainly what you want, but it illustrates
how, from a library perspective, parallel threads create more possible
machine-level concurrency (as parallelism) than just futures.

Blocking regions (that become merely synchronizing) are primarily
identified by `start-atomic` and `end-atomic` pairs. Those atomicity
commands are used within the Racket implementation, and they are also
exposed via `ffi/unsafe/atomic`. That generally the key to this
adaptation. When a parallel thread is running, between calls to those
two functions, the continuation is moved to a coroutine thread with a
(fairly expensive) synchronization. As far as Racket code is
concerned, though, it all runs inside `(current-thread)`.

The first big part of the implementation is making the coroutine
thread scheduler and future thread scheduler play together, and making
things like `kill-thread`, `break-thread`, and `thread-suspend` work
like they should. The second big part is making thr rktio layer have
its own lock, and making individual ports have their own locks. There
are also clocking changes for mutable `eq?`-based and `eqv/`-based
hash tables and pseudo random generators.

Operations that involve a custodian, including opening an dclosing
files, remain blocking, and that's probably the main next area for
improvement.

The recently introduced notion of "uinterruptibe" at the
`ffi/unsafe/atomic` level is made more constrained here. The old idea
was to be flexible and allow a transfer of atomic mode over to a
coroutine thread, since the original use case could be merely a slow
allocator leak if things go wrong. But even a leak would be bad and
unintended, so an error will be better.