extracter: rewrite, bugfix, and test mutated-vars
mutated-vars traverses an expression in a linklet to discover which
variables might be mutated, to enable some optimizations in the
extraction process. Previously, it had one *known* case of silent
unsoundness; however, there were a couple bugs (all in the handling
of let-values and letrec-values) which caused other cases of unsoundness.

This PR rewrites the let-values and letrec-values. First, we addrss all
the bugs (unintentional unsoundness); explicit mutation is now detected
in all cases.

For the implicit mutation, we take a conservative approach. We recognize
some specific obviously-safe instances of letrec-values, specifically,
when the rhs of a letrec is: a plain lambda, a case-lambda, the pattern
generated by a desugared keyword procedure, and non-recursive data. If
every rhs is safe, then no implicit mutations are observable. Otherwise,
we signal an error. (It would be sound to treat them as mutated variables,
See https://github.com/racket/racket/pull/5272 for discussion on why.)

Concretely, not-known-as-safe letrecs occurr very rarely in the actual
code that the extracter is used for. We found 4 instances across three
categories:

1. Two cases where the reference to the letrec variable is inside a
   lambda, but the lambda is not directly bound in the same letrec,
   and the (very basic) analysis can't see that the lambda won't be
   invoked until all bindings are defined. We changed these to use
   explicit mutation.

2. One case where cross-thread synchronization prevents the variable
   from being referenced before its definition. We changed this to
   explicit mutation in the same way.

3. One case which was an actual bug; fixed in b9b0172.

There's an extensive suite of unit tests. Currently, they have to
be run manually with `(require (submod "simplify-defn.rkt" test))`,
but if the build system is modified later to run test submodules,
they'll get run, at least.