cs io, thread: tweaks to appease a small extractor change
The extractor had a bug which could cause unsoundness in how it
handles let-values and letrec-values, and therefore how it
handles any internal definition context. The bug is fixed in
https://github.com/racket/racket/pull/5272, the next commit.
The fix imposes a restriction on code destined for extraction,
and a couple places in trunk run afoul of that restriction.

To be explicit, the code changed here is bug-free; it's just
that the expander isn't smart enough to figure that out.

The specific changes here are to simply make the mutation
explicit (`define` the variable to a placeholder, and then
construct the real value and `set!` it into place) instead
of implicit desugaring into the same thing.

Code by: Matthew Flatt <mflatt@racket-lang.org>
Commit message by: Jade Sailor <jgitpublic.45358@sai9.net>