make uinterruptibe mode more constrained in futures and parallel threads
The old idea was to be flexible and allow a transfer of atomic mode
over to a coroutine thread, since the original use case culd be merely
a slow allocator leak if things go wrong. But even a leak would be bad
and unintended, and the application of uninterruptable mode seems likely
to grow, so an error will be better.