io: add some missing `atomically`s
For rktio functions wrapped to combine a call with gathering an error
(in any) in a vector, an `atomically` wrapper was often missing. I
think I may have intended to put synchronization inside the wrapper in
that case, or maybe I just got confused because it looks like an
atomic call in "src/io". In any case, making the atomicity explicit in
"src/io" will enable further exploration.

The missing `atomically`s are a bug even with just coroutine threads,
because a thread running in between the call and error-vector
allocation could interfere with the reported error, leading to an
incorrect error message. With parallel threads, it's much easier to
provoke bad behavior.