`subprocess-kill` error reporting in CS, work around Mac OS `killpg`
The `subprocess-kill` function on Racket CS was discarding errors, so
fix that.

One error that wasn't reported was an EPERM from Mac OS in the case of
trying to kill a process grup when all processes in the group have
already terminated. That seems like a bug in Mac OS. We work around it
by checking for termination of the main process in the group --- at
the risk of disabling future signalling to other processes in the
group that have not terminated.