rktio: refine Mac OS `killpg` workaround
The previous attempt in 31f324de19 doesn't work, as exposed by CI and
trying more on my local machine. There doesn't seem to be a better
workaround than discarding the error.

Also, repair a broken test in "portlib.rktl". This is an old test, but
parallel threads exposed its problems.