BC: change `processor-count` to always return parallel count
Even when futures and places disabled, `processor-count` now returns
the value that would be returned if futures and places were supported.
This brings BC in line with CS, and it means that `racobc setup` will
use process-based parallelism by default on AArch64 Mac OS. That's
useful, in turn, to avoid timeouts in tests (like
"unix-installer-test.rkt" in "distro-build") that have been
historically run on x86_64 and expect a reasonably fast setup phase.