CS ffi: repair some immobile cell operations
An immobile cell's addresss might flow into a place that claims not to
be GCable and then back to `free-immobile-cell`, which did not handle
those pointers correctly. This became broken with
8c586f93f62c2ceaff01b7017f773abe70fa12a2.

The JPEG encoder ran into this problem, but the main effect is an
extremely slow leak.

Also, clean up locking of struct results from foreign calls, which are
not necessarily allocated as GCable.