thread: repair `custodian-managed-list` concurrent with a GC
Also, fix a problem with mutable hash tables whee keys are removed,
where iteration could become progressively slower. This changes the
behavior of iteration over a mutable hash table where mutation happens
during the iteration; the generate key sequence is not specified in
that case, but this change will tend to make it slow, too.