BC: sync `in-hash` guarantees and refine documented promises
It turns out that BC provided specific guarantees on update and
removal in a mutable hash table only via `hash-map`, and not
`in-hash`. Small changes bring BC in line with CS, and documentation
now explains specifically guarantees for `in-hash`, `in-hash-keys`,
etc.

Another possible direction would have been to weaken the guarantees
that CS and BC had implemented (where neither was a superset of the
other). Since providing `hash-map`-like guarantess is relatively easy,
it seems best to not break programs that were relying before on too
much.