expander: fix kill satefy of registry lock
The namespace registry lock was meant to be released by a thread that
is killed, but that part wasn't implemented correctly if the thread is
not only terminated but also GCed already. The existing test didn't
catch the problem for a long time, but probably because a use of the
lock had gotten lost between the time the test was written and a
recent change to add (or restore?) use in `make-empty-base-namespace`.