CS: repair interaction of semaphores and parallel threads
The CAS-or-atomic combination did not take into account the fact that
a CAS-based post or wait might happen concurrently with an
atomic-based wait or post.

Also, a parallel thread's coroutine component did not atomically
transition to being suspended, so it was possible to be between
releasing the lock on the parallel thread's future and suspending
itself at a point where the future decides whether to reschedule the
coroutine thread.