udp-receive + udp-send-to: address-name cache to reduce name resolution
A common pattern is to use an address and port produced by
`udp-receive` to immediately send via `udp-send-to`. The `udp-recive`
function returns a address string and poirt number, and the
`udp-sent-to` function accepts those back --- but in general that
means sending the address string through a resolution step, which is a
lot slower than just using the address that the low-level receive
operation reports (and even though the address string is always in
numerical format).

The approach here adds a cache that more directly uses the address
from `udp-receive` internally to send back out via `udp-send-to`,
instead of going through a general address-resolution step.

Anther approach would be to introduce new APIs that keep an address
representation more abstract, but then programs have to be adjusted to
take advantage; a simple cache captures the common case automatically.
Yet another possibility is to recognize numerical addresses and take a
shortcut for resolving them, instead going through `getaddrinfo` and
similar; that approach gets some improvement, but the approach here is
still faster.