syntax-local-apply-transformer: support expander inspector
Analogous to the way `local-expand` directly uses a binding's
inspector on the groups that the expander itself is trusted, allow a
caller of `syntax-local-apply-transformer` to supply an expander
inspector alongside a binding that provides the transformer's
inspector.

This extension is needed by Rhombus, for example, to make Rhombus
macros work in a sandboxed environment.