schemify: correct `make-struct-type` handling
Tracking of `make-struct-type-property` guards was not right, leading
to the possibility of `(void)` results in place of a "variable used
before definition" error if a structure-type property guard referenced
a binding too early.

In other cases, tracking was too pessimistic. Instead of treating any
structure-type property with a guard as a reason to consider the whole
property list untrackable for use-before-definition decisions,
pessimize only the value associated with key that is not known to be a
structure-tyoe property without a guard.

More generally, however, guards are too complex for the optimizer to
see through. Add a new `racket/unsafe/struct-type-property` library
with `unsafe-make-struct-type-property/guard-calls-no-arguments`,
which tells the potimizer to make a more optimistic assumption about
property guards. The intent is for this property constructor to be
used for implementing method-like dispatch mechanisms where the
compiler cannot see that methods are gathered and adjusted by a
property's guard, but not called at that point.

Some tests in "optimize.rktl" had been accidentally disabled by a
leftover and broken adjustment in 65864a2bd6, so fix that while
adding new tests.