expander: adjust `syntax-shift-phase-level` to affect label phases
Instead of ignoring integer phase shifts after a shift into the label
phase, have the integer shift affect which bindings are in the label
phase, even after-the-fact.

This change is not intended to affect `require`, where importing with
`(for-label (for-meta X))` is still treated as equivalent to
`(for-label X)`, for example. That's because importing into the label
phase flattens multiple phases into a label phase (and conflicts are
detected and reported), while shifting in general does not flatten.