[expander] improve tests for cross-phase-persistent modules
Summary:

Test slighly stronger invariants when a module is expected to not be
cross-phase-persistent. Instead of just asserting there's some syntax error,
assert (a) there's a syntax error mentioning "cross-phase", (b) no output is
generated, and (c) if an additional expected message is present in the error,
then that it's present.

Test Plan:

Existing tests pass. Using the "additional message" present in the next commit works.