[expander] show which required module breaks cross-phase-persistence
Modules which are declared cross-phase persistent are only permitted
to require other modules if those modules are also cross-phase
persistent. Previously, the expander's error when this check fails was
just:

    module: cannot be cross-phase persistent due to required modules

This is not particularly helpful. This modifies the expander to track
which required module wasn't cross-phase-persistent (or an arbitrarily
chosen such module if multiple required modules weren't cross-phase
persistent) and include it in the error message. For example,

    (module u '#%kernel
      (#%require racket/base)
      (#%declare #:cross-phase-persistent))

triggers the message:

    test.rkt:3:13: module: cannot be cross-phase persistent because some required module isn't
      at: #:cross-phase-persistent
      required module: #<module-path-index:racket/base>

(along with the additional "in", "location" fields; edited out for brevity.)

The error message isn't ideal; it prints the problematic module as a
formatted module-path-index.  On the other hand, cross-phase-persistent
modules are a rare and advanced feature, and some error messages regarding
them already include some implementation gunk (even references to
expander-internal structs sometimes), so I don't think this is a big
deal. (And TBH I was too lazy to figure out the right way to pleasantly
render the mpi.)

Test Plan:

test added (using the support in the previous commit)