BC: adjust old-fashioned `error` convention to match docs and CS
When `error` is called with a string initial argument, according to
the documentation, remaining arguments are supposed to be converted
using `error-value->string-handler`. CS implements that behavior,
probably because it was based on documentation. BC has been formatting
with `write`, instead. Probably this has not shown up as a problem
before, because old-style `error` calls are *really* old style at
thtis point.

Changing the way BC has always wored is not obviously the right
choice, but the documented choice and the one implemented in CS seems
better, so this commit brings them in line by changing BC.