raco pkg: add timeout for network operations
If a server accepts a connection and keeps it open but doesn't respond, then
package-manage operations can get stuck. Add a timeout.

The default timeout is 30 minutes, since a download in a limited user
environment really could take a while without implying a problem, and
users are generally free to hit Ctl-C. A service like the pkg-build
service should configure the timeout as lower.

The custodian-based implementation has the extra benefit of cleaning
up ports that could be left often in certain network-error situations.