thread: repair `sync` in parallel threads
For evts that rely on a cross-place, cross-OS_thread signal for
wake-up, `sync`ing on the evt in a parallel thread could miss the
signal, because it could happen in parallel between the decision that
the evt is not yet ready and blocking the `sync`ing thread.

A reslated problem is that, depending on how fast the OS `select` or
`poll` operation runs, a parellel therad `sync`ing on a blocked evt
could spin instead of eventually sleeping. This tended to happen on
Mac OS and not Linux. When it did happen, it masked the other bug.

The solution for both problems is to keep more of `sync` internally in
a coroutine thread, insteda of bouncing between the couroutine
representive and internal future of a parallel thread. Besides
repairing scheduling, this should also improve performance for `sync`
in a parallel thread.

Related to #5371