RFC: make make-struct-type accept `'current` for an inspector
Summary:

Currently, `make-struct-type` defaults to `(current-inspector)` if no inspector
is specified. However, specifying an inspector is mandatory to specify anything
involving optional args which come after, e.g. that fields should be immutable.
At first glance, this is fine.

However, `make-struct-type` is allowed at the module level of
cross-phase-persistent modules, which have certain restrictions, and
`(current-inspector)` can't be specified explicitly. This leaves authors of
cross-phase-persistent in an unpleasant situation: either have structs be
opaque but always mutable, with no guard, etc.; or have structs which can be
immutable but are transparent.

So, change `make-struct-type` to allow `'current` for an inspector to mean
the current inspector.

There are a couple ways to alternate ways to resolve this tension. Allowing
`(current-inspector)` as a cross-expr in general sounds riskier than modifying
make-struct-type in a backwards-compatible manner. Another would be to ban
non-transparent structs, which seems very undesirable. Finally, we could leave
things inconsistent, which seems sad.

Test Plan:

TBD