remove misleading dead code
The test `(null? #'new-rest)` clearly can never be true, so the
enclosing `if` would never take the "true" branch, which was intended
as an optimization. That branch had rotted a bit, and should have been
`#'unpack` instead of `#'core`. But the ad hoc optimization doesn't
really do anything, since the compiler achieves the same end with
inlining. Meanwhile, changing to enable the optimization attempt does
seem to interfere with Typed Racket (which can be sensitive to
specific expansions). So, we remove the never-true test and unused
branch.

Closes #5383