build: add support for `build/CONFIGURE_ARGS` and `build/PKGS`
Add `build/CONFIGURE_ARGS` as way of supplying `configure` (or
`winfig`) arguments as an alternative to supplying them as each time
with `make CONFIGURE_ARGS=...`. Similarly, `build/PKGS` is an
alternative to `PKGS=""`. This change enables configuring a checkout
so that the choices stick for all future `make`s (or `nmake`s).