`equal?` characters => `eq?`, not just `eqv?` (even on BC)
Update Racket's definition to guarantee the behavior that CS has
always provided, bring BC in line with CS so that we don't need to
worry about this difference anymore.

Due to the way that BC represents characters, guaranteeing only `eqv?`
could provide a performance benefit and reduced memory use, at least
in some cases. The difference is normally small. The worst-case
scenario for this change is a program that has just

```
   (for/fold ([v #f]) ([c (in-range #x110000)]
                       #:unless (<= #xD800 c #xDFFF))
     (integer->char c))
```

where I measure 168 msecs after this change versus 24 msecs before
(versus 2 msecs for Racket CS) on x86_64. It's also possible for
performance to improve as a result of caching character allocations,
and there will be essentially difference in time on a second run of
the above loop.

The implementation here caches characters as they are allocated, and
so it can use up to 40MB or so of extra space per-place (for a program
that uses all characters). The cache table could be made weak, but
that would be slower, and so we don't.

Given that BC is deprecated and generally slow compared to CS, the
benefit of guaranteeing `eq?` on characters seems like the right
choice at this point.