Replace `gen-temp.rkt` with new `make-stx-id-counter` in stx.rkt
Currently, "gen-temp.rkt" exists to provide just `gen-temp-id`, which some
low-level macros use to (surprise) generate IDs for temporaries. This used
a global counter (modulo multiple instantiations of gen-temp.rkt), which
is entirely unnecessary. The only uses of `gen-temp-id` are:

- cond, which can just use a simple single hygenic identifier. (Why was it
using gen-temp-id? I've no idea, but it's been that way as far back as the
svn history goes, as far as I can tell.)

- with-syntax, which uses `gen-temp-id` to bind the evaluation results of the
RHS's of the syntax bindings. These identifiers need to be unique within any
invocation of with-syntax, but standard hygene will take care of everything
beyond that.

- syntax-case, which uses it similarly to `with-syntax`.

We replace `gen-temp-id` with `make-stx-id-counter`, a new function in stx.rkt
which maintains a local counter; `with-syntax` and `syntax-case` use that,
and `cond` just uses a boring identifier.

This allows us to eliminate gen-temp.rkt, which lives at the bottom of the
module hierarchy, and reduces module instantiations by `racket/base` by ~3%.

Test plan: (1) code review the `cond` changes; (2) manually skim the `make
derived` changes, and the only ones are alpha-renaming from the `cond` change
which gives me confidence that with-syntax and syntax-case are still working;
(3) DrRacket runs; (4) CI.